fix: 일반 유저만 삭제하도록 검증 로직 추가

+++ aics-admin/src/main/java/kgu/developers/admin/user/application/UserAdminFacade.java
@@ -27,6 +27,7 @@ public class UserAdminFacade {
 	@Transactional
 	public void kickOutUser(UserKickOutRequest request) {
 		User user = userQueryService.getUserById(request.userId());
+		user.validateDeletable();
 		userCommandService.deleteUser(user);
 	}
